fix rewrites-with-basepath.test.ts
