+++ test/e2e/i18n-ignore-rewrite-source-locale/next.config.js
@@ -1,0 +1,25 @@
+module.exports = {
+  basePath: '/basepath',
+  i18n: {
+    locales: ['en', 'sv', 'nl'],
+    defaultLocale: 'en',
+  },
+  async rewrites() {
+    return {
+      beforeFiles: [
+        {
+          source: '/:locale/rewrite-files/:path*',
+          destination: '/:path*',
+          locale: false,
+        },
+        {
+          source: '/:locale/rewrite-api/:path*',
+          destination: '/api/:path*',
+          locale: false,
+        },
+      ],
+      afterFiles: [],
+      fallback: [],
+    }
+  },
+}
