+++ test/e2e/i18n-ignore-rewrite-source-locale/pages/api/hello.js
@@ -1,0 +1,3 @@
+export default function handler(req, res) {
+  res.send('hello from api')
+}
