+++ test/e2e/i18n-ignore-rewrite-source-locale/public/file.txt
@@ -1,0 +1,1 @@
+hello from file.txt
